+++ work/templates/work/basic_form.html
@@ -10,7 +10,7 @@
 
     <div class="row row--centered">
         <div class="col col--1of2">
-            <form method="post">
+            <form method="post" enctype="multipart/form-data">
                 {% csrf_token %}
 
                 <div class="row-no-margin">
